+++ zed_components/src/tools/include/sl_logging.hpp
@@ -22,6 +22,16 @@
 #define DEBUG_STREAM_COMM(stream_arg) if (mDebugCommon) RCLCPP_DEBUG_STREAM( \
     get_logger(), stream_arg)
 
+// Advanced
+#define DEBUG_ADV(...) if (mDebugAdvanced) RCLCPP_DEBUG(get_logger(), __VA_ARGS__)
+#define DEBUG_ONCE_ADV(...) if (mDebugAdvanced) RCLCPP_DEBUG_ONCE(get_logger(), __VA_ARGS__)
+#define DEBUG_STREAM_ADV(stream_arg) if (mDebugAdvanced) RCLCPP_DEBUG_STREAM( \
+    get_logger(), stream_arg)
+#define DEBUG_STREAM_THROTTLE_ADV(duration, stream_arg) \
+  if (mDebugAdvanced) { \
+    rclcpp::Clock steady_clock(RCL_STEADY_TIME); \
+    RCLCPP_DEBUG_STREAM_THROTTLE(get_logger(), steady_clock, duration, stream_arg);}
+
 // Video Depth
 #define DEBUG_STREAM_VD(stream_arg) if (mDebugVideoDepth) RCLCPP_DEBUG_STREAM( \
     get_logger(), stream_arg)